pyproject.toml
[project]
name = "x-atuo"
version = "0.3.0"
description = "Webhook and scheduled Twitter automation service built on Agent Reach patterns."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn>=0.30",
  "apscheduler>=3.10",
  "langgraph>=0.2.60",
  "langchain>=0.3.0",
  "langchain-core>=0.3",
  "langfuse>=3.0.0",
  "pydantic>=2.8",
  "pydantic-settings>=2.4",
  "pyyaml>=6.0",
  "curl_cffi>=0.7",
  "xclienttransaction>=1.0.1",
  "beautifulsoup4>=4.12",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27",
  "pytest>=8.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/x_atuo"]

[tool.pytest.ini_options]
pythonpath = ["src"]
